Using existing runner for goto definition. Properly handling runner kill.
